+++ framework/CHANGELOG.md
@@ -13,6 +13,19 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 
+## [0.17.1] - 2023-07-26
+
+### Added
+
+- Ability to set admin to native contracts during instantiation
+- Added neutron
+
+### Changed
+
+- Address of App/Adapter returned and set by default.
+
+### Fixed
+
 ## [0.17.0] - 2023-07-05
 
 ### Added